backend/requirements.txt
django==5.1.4
djangorestframework
django-cors-headers
python-decouple
requests
google-generativeai
